hammerhead: rename mpdecision and thermal engine
both we implemented on the kernel side. Due that we want to keep up 3rd
party kernel comptability which may use googles mpdecision and thermal engine
we do not remove them from here. We just rename then. If a third party kernel is
detected an init.d script takes care to name them to the original name back
to turn on the user space implementation.